updates for comments on PR
