+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/domain/IRemindAtNotificationManager.kt
@@ -1,0 +1,14 @@
+package com.realityexpander.tasky.agenda_feature.domain
+
+import android.content.Context
+import android.content.Intent
+
+interface IRemindAtNotificationManager {
+
+    fun showNotification(
+        context: Context,
+        alarmIntent: Intent
+    )
+
+    fun createNotificationChannel(context: Context)
+}
