+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/agenda_screen/AgendaViewModel.kt
@@ -147,10 +147,6 @@ class AgendaViewModel @Inject constructor(
                     //}
                 }
         }
-
-//        viewModelScope.launch {
-//            _oneTimeEvent.tryEmit(OneTimeEvent.SetAllAgendaItemAlarms(agendaItems))
-//        }
     }
 
     fun sendEvent(event: AgendaScreenEvent) {
