+++ app/src/main/java/com/realityexpander/tasky/core/presentation/notifications/RemindAtAlarmManager.kt
@@ -7,6 +7,7 @@ import android.content.Intent
 import android.os.Parcelable
 import com.realityexpander.tasky.MainActivity
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
+import com.realityexpander.tasky.agenda_feature.domain.IRemindAtAlarmManager
 import com.realityexpander.tasky.core.presentation.broadcastReceivers.AlarmBroadcastReceiver
 import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManager.ALARM_NOTIFICATION_INTENT_ACTION_ALARM_TRIGGER
 import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManager.ALARM_NOTIFICATION_INTENT_EXTRA_AGENDA_ITEM
@@ -16,19 +17,6 @@ import com.realityexpander.tasky.core.util.toZonedDateTime
 import logcat.logcat
 import java.time.ZonedDateTime
 import java.time.temporal.ChronoUnit
-
-interface IRemindAtAlarmManager {
-
-    fun setAlarmsForAgendaItems(
-        context: Context,
-        agendaItems: List<AgendaItem>
-    )
-
-    fun cancelAllAlarms(
-        context: Context,
-        onFinished: () -> Unit = {},
-    )
-}
 
 object RemindAtAlarmManager : IRemindAtAlarmManager {
 
@@ -75,7 +63,7 @@ object RemindAtAlarmManager : IRemindAtAlarmManager {
                     CURRENT_ALARMS_TITLES,
                     futureAgendaItems.map { agendaItem ->
                         agendaItem.title
-                    }.toList().toCollection(ArrayList())
+                    }.toCollection(ArrayList())
                 )
             },
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
